+++ internal/tracefuzz/fill.go
@@ -11,6 +11,8 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
+type contextKey struct{}
+
 // Fill sets v to a fuzz-generated value.
 func Fill(f *Fuzzer, v reflect.Value) {
 	if !v.IsValid() {
@@ -71,9 +73,9 @@ func fillStruct(f *Fuzzer, v reflect.Value) {
 		return
 	}
 	if t == reflect.TypeFor[context.Context]() {
-		var ctx context.Context = context.Background()
+		ctx := context.Background()
 		if f.Bool() {
-			ctx = context.WithValue(context.Background(), struct{}{}, f.String())
+			ctx = context.WithValue(context.Background(), contextKey{}, f.String())
 		}
 		v.Set(reflect.ValueOf(ctx))
 
@@ -120,7 +122,7 @@ func fillSpecialPointer(f *Fuzzer, v reflect.Value) {
 func fillSlice(f *Fuzzer, v reflect.Value) {
 	n := f.Intn(4)
 	slice := reflect.MakeSlice(v.Type(), n, n)
-	for i := 0; i < n; i++ {
+	for i := range n {
 		Fill(f, slice.Index(i))
 	}
 	v.Set(slice)
@@ -162,65 +164,6 @@ func fillInterface(f *Fuzzer, v reflect.Value) {
 			return
 		}
 		v.SetZero()
-	}
-}
-
-func typedNil(t reflect.Type) (reflect.Value, bool) {
-	switch t {
-	case reflect.TypeFor[trace.SessionInfo]():
-		var s *fuzzSession
-
-		return reflect.ValueOf((trace.SessionInfo)(s)), true
-	case reflect.TypeFor[trace.TxInfo]():
-		var tx *fuzzTx
-
-		return reflect.ValueOf((trace.TxInfo)(tx)), true
-	case reflect.TypeFor[trace.EndpointInfo]():
-		var e *fuzzEndpoint
-
-		return reflect.ValueOf((trace.EndpointInfo)(e)), true
-	case reflect.TypeFor[trace.ConnState]():
-		var s *fuzzConnState
-
-		return reflect.ValueOf((trace.ConnState)(s)), true
-	case reflect.TypeFor[trace.Call]():
-		var c *fuzzCall
-
-		return reflect.ValueOf((trace.Call)(c)), true
-	case reflect.TypeFor[trace.TableQueryParameters]():
-		var p *fuzzTableQueryParameters
-
-		return reflect.ValueOf((trace.TableQueryParameters)(p)), true
-	case reflect.TypeFor[trace.TableDataQuery]():
-		var q *fuzzTableDataQuery
-
-		return reflect.ValueOf((trace.TableDataQuery)(q)), true
-	case reflect.TypeFor[trace.TableResultErr]():
-		var r *fuzzTableResultErr
-
-		return reflect.ValueOf((trace.TableResultErr)(r)), true
-	case reflect.TypeFor[trace.TableResult]():
-		var r *fuzzTableResult
-
-		return reflect.ValueOf((trace.TableResult)(r)), true
-	case reflect.TypeFor[trace.ScriptingQueryParameters]():
-		var p *fuzzScriptingQueryParameters
-
-		return reflect.ValueOf((trace.ScriptingQueryParameters)(p)), true
-	case reflect.TypeFor[trace.ScriptingResultErr]():
-		var r *fuzzScriptingResultErr
-
-		return reflect.ValueOf((trace.ScriptingResultErr)(r)), true
-	case reflect.TypeFor[trace.ScriptingResult]():
-		var r *fuzzScriptingResult
-
-		return reflect.ValueOf((trace.ScriptingResult)(r)), true
-	case reflect.TypeFor[trace.Issue]():
-		var i *fuzzIssue
-
-		return reflect.ValueOf((trace.Issue)(i)), true
-	default:
-		return reflect.Value{}, false
 	}
 }
