+++ internal/tracefuzz/invoke.go
@@ -64,56 +64,3 @@ func fillTopicInterfaces(f *Fuzzer, v reflect.Value) {
 		}
 	}
 }
-
-func topicInterface(f *Fuzzer, t reflect.Type) (reflect.Value, bool) {
-	s := f.String()
-	name := t.Name()
-	switch name {
-	case "TopicReaderStreamSendCommitMessageStartMessageInfo":
-		switch f.Choice(3) {
-		case 0:
-			return reflect.Value{}, true
-		case 1:
-			var m *fuzzTopicCommitMessage
-
-			return reflect.ValueOf(m), true
-		default:
-			return reflect.ValueOf(&fuzzTopicCommitMessage{commits: nil}), true
-		}
-	case "TopicReaderDataResponseInfo":
-		switch f.Choice(3) {
-		case 0:
-			return reflect.Value{}, true
-		case 1:
-			var r *fuzzTopicDataResponse
-
-			return reflect.ValueOf(r), true
-		default:
-			return reflect.ValueOf(&fuzzTopicDataResponse{bytesSize: f.Intn(100)}), true
-		}
-	case "TopicReadStreamInitRequestInfo":
-		switch f.Choice(3) {
-		case 0:
-			return reflect.Value{}, true
-		case 1:
-			var r *fuzzTopicInitRequest
-
-			return reflect.ValueOf(r), true
-		default:
-			return reflect.ValueOf(&fuzzTopicInitRequest{consumer: s, topics: []string{s}}), true
-		}
-	case "TopicWriterResultMessagesInfoAcks":
-		switch f.Choice(3) {
-		case 0:
-			return reflect.Value{}, true
-		case 1:
-			var a *fuzzTopicWriterAcks
-
-			return reflect.ValueOf(a), true
-		default:
-			return reflect.ValueOf(&fuzzTopicWriterAcks{count: f.Intn(10)}), true
-		}
-	default:
-		return reflect.Value{}, false
-	}
-}
